refactor(dashboard): split 3 largest TS files under 400 lines
Split command-plane.ts (880 lines) into:
- command-plane-core.ts (337): unit/topology/operation/detachment/capacity/alert/trace types
- command-plane-swarm.ts (281): swarm lane/worker/provider/resolution types
- command-plane-orchestra.ts (284): orchestra/chain/help/snapshot types

Split keeper-runtime.ts (715 lines) into:
- keeper-state.ts (391): signals, normalizers, thread state management
- keeper-stream.ts (80): SSE stream event handling and abort logic
- keeper-actions.ts (305): hydrate, send, probe, recover lifecycle actions

Split mission-store.ts (690 lines) into:
- mission-signals.ts (33): signals and poll timer
- mission-normalizers-entities.ts (243): entity-level normalizers
- mission-normalizers.ts (390): operator/aggregate normalizers and top-level transforms
- mission-actions.ts (75): fetch and refresh actions

Original files preserved as barrel re-exports for backward compatibility.
Vite build passes (tsc --noEmit && vite build: ok).
